Use XCTSkip for native binding tests on iOS 16
Replace @available annotations with guard #available + XCTSkip to
prevent XCTest from failing on iOS 16 when it discovers but cannot
execute availability-gated test methods.
