+++ WorkflowSwiftUI/Tests/StoreTests.swift
@@ -772,6 +772,145 @@ final class StoreTests: XCTestCase {
         await fulfillment(of: [countDidChange], timeout: 0)
         XCTAssertEqual(state.count, 1)
     }
+    
+    // MARK: - Native SwiftUI Bindings
+    
+    @MainActor
+    func test_nativeBindings() async throws {
+        guard #available(iOS 17, macOS 14, tvOS 17, watchOS 10, *) else {
+            throw XCTSkip("Requires iOS 17+")
+        }
+
+        var state = State()
+        let model = StateAccessor(state: state) { update in
+            update(&state)
+        }
+        let (_store, _) = Store.make(model: model)
+        @SwiftUI.Bindable var store = _store
+
+        let countDidChange = expectation(description: "count.didChange")
+
+        withObservationTracking {
+            _ = store.count
+        } onChange: {
+            countDidChange.fulfill()
+        }
+
+        let binding = $store.count
+        binding.wrappedValue = 1
+
+        await fulfillment(of: [countDidChange], timeout: 0)
+        XCTAssertEqual(state.count, 1)
+    }
+
+    @MainActor
+    func test_nativeBindingSendingCustomAction() async throws {
+        guard #available(iOS 17, macOS 14, tvOS 17, watchOS 10, *) else {
+            throw XCTSkip("Requires iOS 17+")
+        }
+
+        var state = State()
+        let model = CustomActionModel(
+            accessor: StateAccessor(state: state) { _ in
+                XCTFail("state should not be mutated through accessor")
+            },
+            sink: Sink { action in
+                switch action {
+                case .onCountChanged(let count):
+                    state.count = count
+                case .foo:
+                    XCTFail("unexpected action: \(action)")
+                }
+            }
+        )
+        let (_store, _) = Store.make(model: model)
+        @SwiftUI.Bindable var store = _store
+
+        let countDidChange = expectation(description: "count.didChange")
+
+        withObservationTracking {
+            _ = store.count
+        } onChange: {
+            countDidChange.fulfill()
+        }
+
+        let binding = $store.count.sending(sink: \.sink, action: \.onCountChanged)
+        binding.wrappedValue = 1
+
+        await fulfillment(of: [countDidChange], timeout: 0)
+        XCTAssertEqual(state.count, 1)
+    }
+
+    @MainActor
+    func test_nativeBindingSendingClosure() async throws {
+        guard #available(iOS 17, macOS 14, tvOS 17, watchOS 10, *) else {
+            throw XCTSkip("Requires iOS 17+")
+        }
+
+        var state = State()
+        let model = ClosureModel(
+            accessor: StateAccessor(state: state) { _ in
+                XCTFail("state should not be mutated through accessor")
+            },
+            onCountChanged: { count in
+                state.count = count
+            }
+        )
+        let (_store, _) = Store.make(model: model)
+        @SwiftUI.Bindable var store = _store
+
+        let countDidChange = expectation(description: "count.didChange")
+
+        withObservationTracking {
+            _ = store.count
+        } onChange: {
+            countDidChange.fulfill()
+        }
+
+        let binding = $store.count.sending(closure: \.onCountChanged)
+        binding.wrappedValue = 1
+
+        await fulfillment(of: [countDidChange], timeout: 0)
+        XCTAssertEqual(state.count, 1)
+    }
+
+    @MainActor
+    func test_nativeBindingSendingSingleAction() async throws {
+        guard #available(iOS 17, macOS 14, tvOS 17, watchOS 10, *) else {
+            throw XCTSkip("Requires iOS 17+")
+        }
+
+        var state = State()
+        let model = ActionModel(
+            accessor: StateAccessor(state: state) { _ in
+                XCTFail("state should not be mutated through accessor")
+            },
+            sendAction: Sink<Action> { action in
+                switch action {
+                case .onCountChanged(let count):
+                    state.count = count
+                case .foo:
+                    XCTFail("unexpected action: \(action)")
+                }
+            }.send
+        )
+        let (_store, _) = Store.make(model: model)
+        @SwiftUI.Bindable var store = _store
+
+        let countDidChange = expectation(description: "count.didChange")
+
+        withObservationTracking {
+            _ = store.count
+        } onChange: {
+            countDidChange.fulfill()
+        }
+
+        let binding = $store.count.sending(action: \.onCountChanged)
+        binding.wrappedValue = 1
+
+        await fulfillment(of: [countDidChange], timeout: 0)
+        XCTAssertEqual(state.count, 1)
+    }
 }
 
 @ObservableState
